[Refactor] API 명세가 바뀜에 따라 모델 타입 수정, 로직 수정

+++ OpenMarket/OpenMarket/Sources/Views/ItemCellLabel.swift
@@ -42,7 +42,7 @@ class PriceLabel: ItemCellLabel {
         case normal, discounted
     }
 
-    func setText(by state: State, _ currency: String, _ price: Int) {
+    func setText(by state: State, _ currency: String, _ price: Double) {
         switch state {
         case .normal:
             text = "\(currency) \(price)"
